2025.03.20

+++ JavaSSM/jdbc_sql/exercise01.sql
@@ -3,7 +3,7 @@ USE mybatis_example;
 DROP TABLE USER;
 
 CREATE TABLE `user` (
-  `id` INT(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
   `username` VARCHAR(50) NOT NULL,
   `password` VARCHAR(50) NOT NULL,
   PRIMARY KEY (`id`)
